# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface WorkspaceStats {
  documents: number;
  totalLocations: number;
  current: number;
  former: number;
  planning: number;
  needsReview: number;
}

export interface ProcessingStage {
  stage: string;
  status: string;
  /** @nullable */
  message?: string | null;
  startedAt: string;
  /** @nullable */
  finishedAt: string | null;
}

export type FddSourceRangeStatus = typeof FddSourceRangeStatus[keyof typeof FddSourceRangeStatus];


export const FddSourceRangeStatus = {
  Current: 'Current',
  Former: 'Former',
  Planning: 'Planning',
} as const;

export interface FddSourceRange {
  section: string;
  status: FddSourceRangeStatus;
  /** @nullable */
  exhibit: string | null;
  /** @nullable */
  printedStart: string | null;
  pdfStart: number;
  pdfEnd: number;
}

export type FddExtractionManifestDiscoveryMethod = typeof FddExtractionManifestDiscoveryMethod[keyof typeof FddExtractionManifestDiscoveryMethod];


export const FddExtractionManifestDiscoveryMethod = {
  toc: 'toc',
  'heading-scan': 'heading-scan',
  'legacy-fallback': 'legacy-fallback',
} as const;

export interface FddExtractionManifest {
  discoveryMethod: FddExtractionManifestDiscoveryMethod;
  sourceRanges: FddSourceRange[];
  pagesExamined: number[];
  ocrPages?: number[];
  /**
     * @minimum 0
     * @maximum 1
     * @nullable
     */
  ocrConfidence?: number | null;
  acceptedRows: number;
  rejectedRows: number;
  duplicateRows: number;
  missingAddressRows: number;
  missingContactRows: number;
  addedRows?: number;
  matchedRows?: number;
  updatedRows?: number;
  unchangedRows?: number;
  ambiguousRows?: number;
  collapsedRows?: number;
  removedRows?: number;
  warnings: string[];
  complete: boolean;
}

export interface RecordCounts {[key: string]: number}

export interface FddDocument {
  id: string;
  franchiseName: string;
  /** @nullable */
  fddYear?: number | null;
  filename: string;
  uploadDate: string;
  processingStatus: string;
  pageCount: number;
  /** @nullable */
  dataAsOf?: string | null;
  /** @nullable */
  sourceExhibit?: string | null;
  extractionManifest: FddExtractionManifest | null;
  /** @nullable */
  lastProcessedAt: string | null;
  locationCount: number;
  recordCounts: RecordCounts;
  stages: ProcessingStage[];
}

export interface DocumentImportInput {
  /** @minLength 1 */
  filename: string;
  /** @minLength 1 */
  objectPath: string;
}

export interface UploadMetadata {
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 1
     * @maximum 104857600
     */
  size: number;
  contentType: string;
}

export interface UploadDestination {
  uploadUrl: string;
  objectPath: string;
}

export interface Contact {
  /** @nullable */
  rawName: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  middleName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  suffix: string | null;
  /** @nullable */
  title: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  phone: string | null;
}

export type FranchiseLocationStatus = typeof FranchiseLocationStatus[keyof typeof FranchiseLocationStatus];


export const FranchiseLocationStatus = {
  Current: 'Current',
  Former: 'Former',
  Planning: 'Planning',
} as const;

export type FranchiseLocationReviewStatus = typeof FranchiseLocationReviewStatus[keyof typeof FranchiseLocationReviewStatus];


export const FranchiseLocationReviewStatus = {
  Needs_review: 'Needs review',
  Approved: 'Approved',
  Rejected: 'Rejected',
} as const;

export interface FranchiseLocation {
  id: string;
  documentId: string;
  /** @nullable */
  sourceDocumentFilename?: string | null;
  franchiseName: string;
  franchisor: string;
  status: FranchiseLocationStatus;
  /** @nullable */
  franchiseeEntity?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zip?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  locationCode?: string | null;
  /** @nullable */
  exitReason?: string | null;
  /** @nullable */
  sourcePage?: number | null;
  /** @nullable */
  printedPage?: string | null;
  /** @nullable */
  sourceExhibit?: string | null;
  /** @nullable */
  sourceSection?: string | null;
  /** @nullable */
  rawSourceText?: string | null;
  confidence: number;
  reviewStatus: FranchiseLocationReviewStatus;
  /** @nullable */
  reviewReason?: string | null;
  contacts: Contact[];
}

export type LocationUpdateReviewStatus = typeof LocationUpdateReviewStatus[keyof typeof LocationUpdateReviewStatus];


export const LocationUpdateReviewStatus = {
  Needs_review: 'Needs review',
  Approved: 'Approved',
  Rejected: 'Rejected',
} as const;

export interface LocationUpdate {
  franchisor?: string;
  /** @nullable */
  franchiseeEntity?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zip?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  exitReason?: string | null;
  reviewStatus?: LocationUpdateReviewStatus;
}

export type ListLocationsParams = {
q?: string;
status?: ListLocationsStatus;
franchisor?: string;
state?: string;
documentId?: string;
reviewStatus?: ListLocationsReviewStatus;
/**
 * @minimum 1
 * @maximum 500
 */
limit?: number;
/**
 * @minimum 0
 */
offset?: number;
};

export type ListLocationsStatus = typeof ListLocationsStatus[keyof typeof ListLocationsStatus];


export const ListLocationsStatus = {
  Current: 'Current',
  Former: 'Former',
  Planning: 'Planning',
} as const;

export type ListLocationsReviewStatus = typeof ListLocationsReviewStatus[keyof typeof ListLocationsReviewStatus];


export const ListLocationsReviewStatus = {
  Needs_review: 'Needs review',
  Approved: 'Approved',
  Rejected: 'Rejected',
} as const;