# lib/api-zod/src/generated/types/fddExtractionManifest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FddExtractionManifestDiscoveryMethod } from './fddExtractionManifestDiscoveryMethod';
import type { FddSourceRange } from './fddSourceRange';

export interface FddExtractionManifest {
  discoveryMethod: FddExtractionManifestDiscoveryMethod;
  sourceRanges: FddSourceRange[];
  pagesExamined: number[];
  ocrPages?: number[];
  /**
     * @minimum 0
     * @maximum 1
     * @nullable
     */
  ocrConfidence?: number | null;
  acceptedRows: number;
  rejectedRows: number;
  duplicateRows: number;
  missingAddressRows: number;
  missingContactRows: number;
  addedRows?: number;
  matchedRows?: number;
  updatedRows?: number;
  unchangedRows?: number;
  ambiguousRows?: number;
  collapsedRows?: number;
  removedRows?: number;
  warnings: string[];
  complete: boolean;
}